backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database (psycopg3 for Python 3.13 support)
sqlmodel==0.0.22
sqlalchemy>=2.0.0,<2.1.0
psycopg[binary]>=3.1.0
alembic>=1.13.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# AI
openai>=1.12.0

# Email & Scheduling
resend>=2.0.0
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
slowapi>=0.1.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
